requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.4
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
ghp-import==2.1.0
idna==3.6
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.5.3
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
postgres==4.0
psycopg==3.1.18
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-openid==2.2.5
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
watchdog==4.0.0
